% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{set.constraint1}
\alias{set.constraint1}
\title{Position limits and Short-Selling constraints}
\usage{
set.constraint1(constraints, times, prices, N, X)
}
\value{
constraint1
}
\description{
Description
}
