% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.constraint2}
\alias{set.constraint2}
\title{Zero-Position Assumption}
\usage{
set.constraint2(constraints, X)
}
\arguments{
\item{constraints}{constraints}

\item{X}{The trade volume in our strategies at price P.}
}
\value{
constraint2
}
\description{
After the conclusion of a trading period, we should not hold any non-zero position.
}
