% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.constraint3}
\alias{set.constraint3}
\title{Market Liquidity Assumption}
\usage{
set.constraint3(constraints, X, V)
}
\arguments{
\item{constraints}{constraints}

\item{X}{The trade volume in our strategies at price P.}

\item{V}{The actual trade volume corresponding to the price P.}
}
\value{
constraint3
}
\description{
At any given moment and price, regardless of whether we are buying or selling, the absolute value of our trade volume should not exceed the actual trading volume at that moment and price.
}
