% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.constraint4}
\alias{set.constraint4}
\title{Modified constraint for constraint1.}
\usage{
set.constraint4(constraints, times, prices, N, X)
}
\arguments{
\item{constraints}{1}

\item{times}{1}

\item{prices}{1}

\item{N}{1}

\item{X}{1}
}
\value{
constraint4
}
\description{
Description
}
