% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.constraints}
\alias{set.constraints}
\title{Constraints}
\usage{
set.constraints(data, N, X, V, short_sell = short_sell)
}
\arguments{
\item{data}{A trading-level dataset should include the following columns: “time,” which represents the trade time of a transaction (in seconds); “price,” which indicates the transaction price; and “volume,” which denotes the transaction volume.}

\item{N}{The maximum number of shares to hold.}

\item{X}{The trade volume in our strategies at price P.}

\item{V}{The actual trade volume corresponding to the price P.}

\item{short_sell}{represents whether short selling is allowed.}
}
\value{
All the constraints
}
\description{
The constraints associated with the aforementioned optimization problem.
}
