% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Problem.R
\name{set.problem}
\alias{set.problem}
\title{Set problem}
\usage{
set.problem(data, N, short_sell = FALSE, tcr = 0)
}
\arguments{
\item{data}{}
}
\value{
The final data and minimum value
}
\description{
Set problem
}
