% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.constraint1}
\alias{set.constraint1}
\title{Position limits and Short-Selling constraints}
\usage{
set.constraint1(constraints, times, prices, N, X)
}
\arguments{
\item{constraints}{constraints}

\item{times}{times}

\item{prices}{prices}

\item{N}{The maximum number of shares to hold.}

\item{X}{The trade volume in our strategies at price P.}
}
\value{
constraint1
}
\description{
In each period, our position cannot exceed the maximum position size N.
}
