% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.constraint1}
\alias{set.constraint1}
\title{Position limits and Short-Selling constraints}
\usage{
set.constraint1(constraints, times, prices, N, X)
}
\arguments{
\item{constraints}{1}

\item{times}{1}

\item{prices}{1}

\item{N}{1}

\item{X}{1}
}
\value{
constraint1
}
\description{
Description
}
