% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.constraint2}
\alias{set.constraint2}
\title{Zero-Position Assumption}
\usage{
set.constraint2(constraints, X)
}
\arguments{
\item{constraints}{1}

\item{X}{1}
}
\value{
constraint2
}
\description{
Description
}
