% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.constraint4}
\alias{set.constraint4}
\title{Modified constraint for constraint1.}
\usage{
set.constraint4(constraints, times, prices, N, X)
}
\arguments{
\item{constraints}{constraints}

\item{times}{times}

\item{prices}{prices}

\item{N}{The maximum number of shares to hold.}

\item{X}{The trade volume in our strategies at price P.}
}
\value{
constraint4
}
\description{
Remove the short-selling constraint, i.e., allowing negative positions.
}
