% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.constraints}
\alias{set.constraints}
\title{Set all constraints}
\usage{
set.constraints(data, N, X, V, short_sell = short_sell)
}
\arguments{
\item{data}{1}

\item{N}{1}

\item{X}{1}

\item{V}{1}

\item{short_sell}{1}
}
\value{
All constraints
}
\description{
Description
}
