% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Position.R
\name{set.problem}
\alias{set.problem}
\title{Set problem}
\usage{
set.problem(data, N, short_sell = FALSE, tcr = 0)
}
\arguments{
\item{data}{1}

\item{N}{1}

\item{short_sell}{1}

\item{tcr}{1}
}
\value{
The final data and minimum value
}
\description{
Set problem
}
